merge develop into main (3nd release)

+++ src/main/java/com/example/barrier_free/domain/facility/repository/FacilityRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.barrier_free.domain.facility.entity.Facility;
 
-public interface FacilityRepository extends JpaRepository<Facility, Long> {
+public interface FacilityRepository extends JpaRepository<Facility, Integer> {
 
 	List<Facility> findAllByIdIn(List<Integer> ids);
 }
